test(runner): run the websocket close tests on Windows too
WebsocketClient.Close has no Unix-specific behavior and the backhaul client
runs on Windows, but client_close_test.go carried //go:build !windows and so
never ran there. The tag was never about the code under test: the file needed
trackedConn and newWshServer, and both lived in pty_recovery_test.go, which is
!windows because of the pty.

Inverting that dependency clears it. trackedConn and dialTracked move to an
untagged conn_test.go, which pty_recovery_test.go can still use because a
tagged file may reference symbols from an untagged one in the same package.
The failure test then drops newWshServer for newCloseReplyServer: the Websh
double registers a pty recovery route and mutates the global ServerURL, and
with failWrites on nothing reaches the peer anyway, so none of that was doing
any work.

build-and-test.yml states the policy this restores: Unix-only packages are
gated deliberately, never by a silent blanket skip. GOOS=windows go list now
shows client_close_test.go and conn_test.go in the package's test files.

Constraint: pty_recovery_test.go stays !windows, so shared helpers cannot live there
Confidence: high
Scope-risk: narrow
Reversibility: clean
