+++ pkg/runner/pty_recovery_test.go
@@ -6,8 +6,6 @@ package runner
 
 import (
 	"context"
-	"errors"
-	"net"
 	"net/http"
 	"net/http/httptest"
 	"runtime"
@@ -22,31 +20,6 @@ import (
 	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/require"
 )
-
-// trackedConn fails writes on demand, so a test can play out a broken connection without breaking a real one.
-type trackedConn struct {
-	net.Conn
-	failWrites    atomic.Bool
-	closed        atomic.Bool
-	readDeadlines atomic.Int32
-}
-
-func (c *trackedConn) SetReadDeadline(t time.Time) error {
-	c.readDeadlines.Add(1)
-	return c.Conn.SetReadDeadline(t)
-}
-
-func (c *trackedConn) Write(b []byte) (int, error) {
-	if c.failWrites.Load() {
-		return 0, errors.New("simulated write failure")
-	}
-	return c.Conn.Write(b)
-}
-
-func (c *trackedConn) Close() error {
-	c.closed.Store(true)
-	return c.Conn.Close()
-}
 
 // wshServer is a local Websh test double: a WebSocket endpoint plus the pty-channels recovery API, with server-side handles to kill connections.
 type wshServer struct {
